replace exp3p with bandit version
